fixes
1) Сделал рефакторинг кода
2) Добавил валидацию для favoriteId
