+++ controllers/drinks/index.js
@@ -1,7 +1,6 @@
 const { ctrlWrapper } = require("../../services");
 
 const getPopularDrinks = require("./popular");
-
 module.exports = {
   getPopularDrinks: ctrlWrapper(getPopularDrinks),
 };
