+++ routes/api/drinks.js
@@ -1,14 +1,9 @@
 const express = require("express");
-const {
-  getFav,
-  addFaV,
-  removeFav,
-} = require("../../controllers/drinks/favorites");
 const { getPopularDrinks } = require("../../controllers/drinks/index");
 const validateBody = require("../../middlewares/validateBody");
-
+const schema = require("../../schemas/JoiValidator");
 const router = express.Router();
-
+const fav = require("../../controllers/drinks/favorite/index");
 const ctrl = require("../../controllers/drinks/drinks");
 const authentication = require("../../middlewares/authentication");
 
@@ -20,9 +15,19 @@ router.get("/own", authentication, ctrl.getOwnDrinks);
 router.post("/own/add", authentication, ctrl.addOwnDrink);
 router.delete("/own/remove/:id", authentication, ctrl.removeOwnDrink);
 
-router.get("/favorite", authentication, getFav);
-router.post("/favorite/add", authentication, addFaV);
-router.delete("/favorite/remove", authentication, removeFav);
+router.get("/favorite", authentication, fav.getFav);
+router.post(
+  "/favorite/add",
+  authentication,
+  validateBody(schema.schemaFavoriteId),
+  fav.addFav
+);
+router.delete(
+  "/favorite/remove",
+  authentication,
+  validateBody(schema.schemaFavoriteId),
+  fav.removeFav
+);
 
 router.get("/:id", authentication, ctrl.getDrinkById);
